First pass completion of API calls JSON write
Built out Transfer and TransferCollection classes which talk hold
information about ambulance pickups. Also built out Hospital and
HospitalCollection which deals with hospitals. Have developed a csv
scraper which extracts relevant information and populates these classes.
The use this info to make a call to an API, process results and return
it in a csv or json format.